src/calendar/service.py
"""
Google Calendar service.

Handles CRUD operations for calendar events.
"""
from datetime import datetime
from typing import Optional

from google.oauth2.credentials import Credentials
from google.auth.transport.requests import Request
from googleapiclient.discovery import build
import structlog

from src.core.config import settings
from src.calendar.colors import get_color_for_status

logger = structlog.get_logger(__name__)


class CalendarService:
    """Google Calendar service wrapper."""

    def __init__(self):
        self.credentials = self._load_credentials()
        self.service = build("calendar", "v3", credentials=self.credentials)

    def _load_credentials(self) -> Credentials:
        """
        Load Google Calendar OAuth2 credentials.

        Expects token.json file in project root.
        """
        # TODO: Implement in Phase 3
        pass

    async def create_event(
        self,
        summary: str,
        start_time: datetime,
        end_time: datetime,
        description: Optional[str] = None,
        status: str = "PENDING"
    ) -> str:
        """
        Create calendar event.

        Args:
            summary: Event title (e.g., "Cita - Dr. González")
            start_time: Event start time
            end_time: Event end time
            description: Event description
            status: Appointment status for color

        Returns:
            Event ID
        """
        # TODO: Implement in Phase 3
        pass

    async def update_event_color(self, event_id: str, status: str) -> None:
        """
        Update calendar event color based on status.

        Args:
            event_id: Google Calendar event ID
            status: New appointment status
        """
        color_id = get_color_for_status(status)

        # TODO: Implement in Phase 3
        logger.info(
            "updating_calendar_event",
            event_id=event_id,
            status=status,
            color_id=color_id
        )

    async def delete_event(self, event_id: str) -> None:
        """
        Delete calendar event.

        Args:
            event_id: Google Calendar event ID
        """
        # TODO: Implement in Phase 3
        pass
