requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis[hiredis]==5.0.1

# External APIs
groq==0.4.2
twilio==8.11.1
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Scheduler
apscheduler==3.10.4

# Logging & Monitoring
structlog==24.1.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utils
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0
